Move visitors.go into runtime/
